% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DraftScrapewithProjections.R
\name{DraftScrapewithProjections}
\alias{DraftScrapewithProjections}
\title{Scraping Projections from variety of sources for FF Draft}
\usage{
DraftScrapewithProjections(season)
}
\arguments{
\item{season}{Specific NFL season}
}
\description{
Scraping Projections from variety of sources for FF Draft
}
\examples{
Draftprojections<-DraftScrapewithProjections(2022)

}
