% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{moveMe}
\alias{moveMe}
\title{Moveme - Move columns}
\usage{
moveMe(data, tomove, where = "last", ba = NULL)
}
\arguments{
\item{data}{Dataframe.}

\item{tomove}{Specific Column(s).}

\item{where}{Location of Column relative to others.}
}
\description{
Moveme - Move columns
}
\examples{
ProjectionsFinal<-moveMe(ProjectionsFinal, c("Season", "Week"), "first")
ProjectionsFinal<-moveMe(ProjectionsFinal, c("Season", "Week"), "last")
}
