% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrapewithProjections.R
\name{ScrapewithProjections}
\alias{ScrapewithProjections}
\title{Scraping Projections from variety of sources}
\usage{
ScrapewithProjections(season, week)
}
\arguments{
\item{season}{Specific NFL season}

\item{week}{Specific Week within NFL Season}
}
\description{
Scraping Projections from variety of sources
}
\examples{
Projections<-ScrapewithProjections(2021, 15)

}
