% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statsfunction.R
\name{Statsfunction}
\alias{Statsfunction}
\title{Stats during NFL season from nflVerse Packages}
\usage{
Statsfunction(
  seasons,
  stat_type = c("players", "defense_by_position", "defense_by_depthposition", "teams")
)
}
\arguments{
\item{seasons}{Specific NFL season}

\item{stat_type}{Type of Stats ("players", "defense_by_position","defense_by_depthposition","teams")}
}
\description{
Stats during NFL season from nflVerse Packages
}
\examples{
PlayerStats<-Statsfunction(Seasonforinputs,"players")


}
