% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanProjections.R
\name{CleanProjections}
\alias{CleanProjections}
\title{Removing many columns from initial scraping.
Must use "DraftScrapewithProjections" prior to this}
\usage{
CleanProjections(DraftProjections)
}
\arguments{
\item{DraftProjections}{An output from the \link{DraftScrapewithProjections} function}
}
\description{
Removing many columns from initial scraping.
Must use "DraftScrapewithProjections" prior to this
}
